ai_service/requirements.txt
flask
flask-cors
pillow
numpy
opencv-python
